Move hyperion logs to somewhere common and generalise get_logging_dir
